auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authentication service"
authors = ["Ihar Shynkarou <ishynkarou@remedypointsolutions.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = {version = "~=1.4", optional = true}
ariadne = "~=0.11"
bcrypt = "~=3.1"
common = { path = "../common/" }
marshmallow = "~=3.7"
psycopg2-binary = {version = "~=2.8", optional = true}
pyjwt = "~=1.7"
uvicorn = "~=0.11"
starlette = "~=0.13"

[tool.poetry.dev-dependencies]
black = "19.10b0"
flake8 = "~=3.7"
flake8-bugbear = "~=20.1"
flake8-comprehensions = "~=3.2"
isort = "~=4.3"
pytest = "~=5.3"
pytest-cov = "~=2.10"
seed-isort-config = "~=1.9"

[tool.poetry.extras]
migrations = ["alembic", "psycopg2-binary"]

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
balanced_wrapping = true
combine_as_imports = true
combine_star = true
include_trailing_comma = true
intend = 4
known_third_party = []
lines_between_types = true
line_length = 119
multi_line_output = 3
order_by_type = true
use_parentheses = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
